scrapytutorial/requirements.txt
scrapy
scrapy-playwright
python-dotenv
# playwright install
pandas